[210]-Rental-Refactoring
1.Product 리팩토링
- Member <-> ProductBoard 관계 적용
- email <-> sellerId @ManyToOne 관계를 적용함에 따라 내용 리팩토링
- 주요부분 ProductService - sellerId 부분

2. GetProducts 리스트 불러오는 부분 (ProductSearch, ProductService)
- 이슈 : 어제 테스트시 필터가 비정상 동작함
- 원인 : RentalHistory - product_id 관계를 삭제해놨었는데 여기서 문제가 됨, RentalHistory에서 returnDate를 가져오는 부분에서 에러가 발생
- 관계를 ManyToOne을 해놨으니까 나중에 한번 더 리팩토링 필요

3. CorsConfig 추가

+++ src/test/java/com/bom/rentalmarket/product/service/ProductGetServiceTest.java
@@ -34,7 +34,7 @@ public class ProductGetServiceTest {
     int pageSize = 10;
 
     List<GetProductForm> productList = productService.getProducts(categoryType, statusType, keyword,
-        pageNo, pageSize, userRegion);
+        (long) pageNo, (long) pageSize, userRegion);
 
     // then
     assertNotNull(productList); // products map 객체가 null이 아닌지 확인
